refactor(native): split proton config and state
